Display total # of licenses & releases per status
Introduce two new functions, count_keys() and count_releases() to grab
an aggregate count of how many keys or releases are in a particualar
status. For now, these values aren't cached. But caching should
probably be introduced at a future point.

In the Licenses and Releases list tables, utilize these functions to
display the total count of each model in each status grouping.