Refetch product object to get updated date values
In 3e42ad6 all instances of it_exchange_get_product_feature were
replaced with calls on the \ITELIC\Product object. As a result, in
WP-CLI we had to fetch the product earlier in the create_product
method. This meant that when we updated the product's date, the prouct
model was out of sync with the database values.

This commit refreshes the model's data by recalling itelic_get_product.