Display total license & release count in list table
In 92c69d1 we added the ability to display the total number of releases
and licenes with a particular status. This did not display the total
for the 'All' option, nor were the counts displayed when a single
status was selected.

Include the 'All' option when counting models, and ensure that the
counts ae displayed, even when the status is selected.