Prevent activating non-active license keys
In the admin manage license screen, disable the 'Remote Activate'
section if the key is disabled or expired. If the key's status is
changed back to active, re-enable the section using javascript. If the
key is originally active, and its status is changed to expired or
disabled, disable the section.

Additionally, in the front-end licenses manager, prevent activating
a key that is not active.