Update download name when releasing a new version
When activating a release, if the name of the product download is a
version of the filename, update the name to be a version of the new
download file. If not, leave the download name alone.

For example, if the download is named 'Classes for iThemes Exchange'
when we release version 1.1 of the download, we would keep the download
name the same. However, if the download is named
'exchange-addon-classes-1.0.zip', we would rename the file to be
'exchange-addon-classes-1.1.zip'.