Load API on parse_request for performance
Currently, the API is loaded on template_redirect. This means that the
main WP_Query has already been run. Additionally, the hook is simply
later in the load process meaning more code has had a chance to run.

Change the hook the API is loaded on to parse_request. This has the
benefits of occurring earlier in the load process, and the rewrite
rules have already been evaluated. This means Dispatch::process() now
accepts a `WP` object instead of a `WP_Query` object.

Updates dispatch unit tests as well.