Add support for guest checkout
Add support for generating keys for guest transactions. Unfortunately,
Exchange's guest checkout functionality does not consistently return
IT_Exchange_Customer objects for the transaction customer, this
requires us to manually instantiate customer objects in a number of
places.

Guest keys won't have a customer ID stored in the database. Instead,
when needed to retrieve the customer, we retrieve it from the
transaction. This means that there is no simple way of retrieving all
license keys by email. But this is not too concerning to me. That would
be a possibility if Exchange created dummy accounts in the background.
But without this, each transaction should be treated as completely
independent. Even if the same email address is used.