Support sending renewal reminders to guests
Introduce new class `Guest_Notification` that is composed with a
regular notification object. This adds support for sending renewal
reminders to customers who have checked out as guest.

This relies on Exchange's built in functions for guest customers, which
are not available if guest checkout is disabled at a later point. This
will require a fix in Exchange.

Adds unit tests.